Edit README.md
Wrote a description of the project including a
list of the skills I plan to demonstrate. This
gives context to the repository, serves as a plan
to guide the outlining and coding stages, and will
allow for a post-project reflection on how the
final project compares to its inital design.